requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-extensions==3.2.3
django-filter==24.3
django-js-asset==2.2.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
